escriba-medico-dev/backend/requirements.txt
fastapi[all]
uvicorn[standard]
python-dotenv
openai
deepgram-sdk
python-multipart
